+++ data/law_dept_sample/services.json
@@ -1,1 +1,8 @@
-["Appeals", "FOIA", "Lobbying"]
+[
+    { "id" : "Appeals",
+      "name" : "Appeals"},
+    { "id" : "FOIA",
+      "name" : "FOIA" },
+    { "id" : "Lobbying",
+      "name" : "Lobbying"}
+]
