+++ js/init.js
@@ -5,22 +5,15 @@ import { loadRevenuePage } from './pages/03_revenue/main.js'
 import { loadPageState } from './utils/storage-handlers.js'
 import { initializeNavButtons } from './components/nav_buttons/nav_buttons.js';
 import { loadPersonnelPage } from './pages/04_personnel/main.js';
+import { addTarget } from './components/sidebar/sidebar.js';
 
-// running tallies of total spend
-let personnel_supp = 0;
-let personnel_baseline = 0;
-let nonpersonnel_supp = 0;
-let nonpersonnel_baseline = 0;
-let target = 2000000;
-let baseline_revenue = 0;
-let supp_revenue = 0;
-let supp_total = personnel_supp - supp_revenue;
-let baseline_total = personnel_baseline - baseline_revenue;
+export let DATA_ROOT = '../../../data/law_dept_sample/'
 
 document.addEventListener('DOMContentLoaded', function () {
 
     var page_state = loadPageState();
     initializeNavButtons();
+    addTarget(2000000);
 
     switch (page_state){
         case 'welcome':
