+++ js/pages/04_personnel/rollup-helpers.js
@@ -27,55 +27,6 @@ function exitAccountEditModal() {
     document.getElementById('editing').removeAttribute('id');
 }
 
-// update sidebar and also cost totals when the FTEs are edited
-function updateDisplayandTotals(rowToEdit){
-
-    let avg_salary = rows[i].querySelector('.salary').getAttribute('value');
-    let baseline_ftes_cell = rows[i].querySelector('.ftes-baseline');
-    let baseline_ftes = baseline_ftes_cell.getAttribute('value');
-    let supp_ftes_cell = rows[i].querySelector('.ftes-supp');
-    let supp_ftes = supp_ftes_cell.getAttribute('value');
-
-
-    // calculate for each row
-    let rows = document.getElementsByTagName('tr');
-    for (let i = 1; i < rows.length; i++){
-        // get all the right values
-        let avg_salary = rows[i].querySelector('.salary').getAttribute('value');
-        let baseline_ftes_cell = rows[i].querySelector('.ftes-baseline');
-        let baseline_ftes = baseline_ftes_cell.getAttribute('value');
-        let supp_ftes_cell = rows[i].querySelector('.ftes-supp');
-        let supp_ftes = supp_ftes_cell.getAttribute('value');
-        // calcuate #FTEs x average salary + COLA adjustments + merit adjustments + fringe
-        total_baseline_cost = baseline_ftes * avg_salary * (1 + fringe) * (1 + cola) * (1 + merit);
-        total_supp_cost = supp_ftes * avg_salary * (1 + fringe) * (1 + cola) * (1 + merit);
-       
-        // update counters
-        personnel_baseline += total_baseline_cost;
-        personnel_supp += total_supp_cost;
-
-        // update totals in table
-        rows[i].querySelector('.calculated-total-supp').textContent = formatCurrency(total_supp_cost);
-        rows[i].querySelector('.calculated-total-baseline').textContent = formatCurrency(total_baseline_cost);
-        
-        // actions if there is an update
-        if ((baseline_ftes + supp_ftes) > 0){
-            // make ? icon visible 
-            let buttons = rows[i].querySelectorAll('.icon-button.btn-see-calcs');
-            buttons.forEach(function(button) {
-                button.style.display = "block";
-            });
-                // update colors if relevant
-            if (baseline_ftes > 0){
-                baseline_ftes_cell.classList.add("keep");
-            } 
-            if (supp_ftes > 0){
-                supp_ftes_cell.classList.add("supp");
-            } 
-        }
-    }
-    updateDisplay();
-}
 
 // check if all service boxes are filled
 function validateServiceSelections(){
